Fix all service modules: DB alignment, entity fixes, STM actions, BDD tests
Modules fixed (all unit/mapper tests pass):
- Supplier (75), Cart (83), Product (74), Promo (23), Notification (39)
- Checkout (11), User (42), Inventory (42), Offer (27), Shipping (52)
- Review (16), ReturnRequest (39), Order (7), Settlement (7)
- Onboarding (27), Supplier-Lifecycle (15), Rules-Engine (47)
- Compliance (45), Support (27), Fulfillment (14)
- Reconciliation (16, new domain+infrastructure modules created)
- Return-Processing (16)

Common fixes across modules:
- DB migration is source of truth — entity/model fields aligned to match
- Missing getActivities() getter added (Jackson serialization)
- Mapper mergeEntity() method added for optimistic locking
- Null-safe version handling (Long->long NPE prevention)
- Money API: Money.of(long,String) not new Money(BigDecimal,String)
- ChenileJpaEntityStore replaces custom stores (ID propagation fix)
- Removed @Component/@Service/@Repository (all @Bean in Configuration)
- Missing STM transition actions and DTOs created
- OWIZ bean name mismatches fixed
- Test data aligned with DB schema columns

BDD tests: Keycloak Testcontainer timeout on local machine (Docker resource limit).
All modules compile and unit tests pass. BDD tests need CI environment with more Docker resources.

+++ promo/promo-domain/src/main/java/com/homebase/ecom/promo/rule/ConditionEvaluator.java
@@ -2,7 +2,6 @@ package com.homebase.ecom.promo.rule;
 
 import com.homebase.ecom.promo.model.CartSnapshot;
 import com.homebase.ecom.promo.model.Operator;
-import com.homebase.ecom.shared.Money;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -42,7 +41,7 @@ public class ConditionEvaluator {
     }
 
     private boolean evaluateAmountThreshold(Condition condition, CartSnapshot cart) {
-        BigDecimal totalAmount = cart.getTotalAmount().getAmount();
+        long totalAmount = cart.getTotalAmount().getAmount();
         return applyOperator(totalAmount, condition.getOperator(), condition.getValue());
     }
